move heap tests out to flat folder
