+++ tests/library/dbg/tests/test_find_fake_fast.py
@@ -4,10 +4,10 @@ import re
 
 import pytest
 
-from .....host import Controller
-from .. import get_binary
-from .. import launch_to
-from .. import pwndbg_test
+from ....host import Controller
+from . import get_binary
+from . import launch_to
+from . import pwndbg_test
 
 HEAP_FIND_FAKE_FAST = get_binary("heap_find_fake_fast.native.out")
