+++ tests/library/dbg/tests/test_heap_bins.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from .....host import Controller
-from .. import get_binary
-from .. import launch_to
-from .. import pwndbg_test
+from ....host import Controller
+from . import get_binary
+from . import launch_to
+from . import pwndbg_test
 
 BINARY = get_binary("heap_bins.native.out")
